+++ x/crosschain/keeper/msg_server_vote_inbound_tx_test.go
@@ -10,6 +10,8 @@ import (
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/evmos/ethermint/x/evm/statedb"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/zeta-chain/zetacore/pkg/chains"
@@ -45,7 +47,7 @@ func setObservers(t *testing.T, k *keeper.Keeper, ctx sdk.Context, zk keepertest
 // https://github.com/zeta-chain/node/issues/1542
 func TestKeeper_VoteOnObservedInboundTx(t *testing.T) {
 	t.Run("successfully vote on evm deposit", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
+		k, ctx, sdkk, zk := keepertest.CrosschainKeeper(t)
 		msgServer := keeper.NewMsgServerImpl(*k)
 		validatorList := setObservers(t, k, ctx, zk)
 		to, from := int64(1337), int64(101)
@@ -61,6 +63,13 @@ func TestKeeper_VoteOnObservedInboundTx(t *testing.T) {
 		zk.ObserverKeeper.SetTSS(ctx, sample.Tss())
 
 		msg := sample.InboundVote(0, from, to)
+
+		err := sdkk.EvmKeeper.SetAccount(ctx, ethcommon.HexToAddress(msg.Receiver), statedb.Account{
+			Nonce:    0,
+			Balance:  big.NewInt(0),
+			CodeHash: crypto.Keccak256(nil),
+		})
+		require.NoError(t, err)
 		for _, validatorAddr := range validatorList {
 			msg.Creator = validatorAddr
 			_, err := msgServer.VoteOnObservedInboundTx(
@@ -78,7 +87,7 @@ func TestKeeper_VoteOnObservedInboundTx(t *testing.T) {
 		require.Equal(t, ballot.BallotStatus, observertypes.BallotStatus_BallotFinalized_SuccessObservation)
 		cctx, found := k.GetCrossChainTx(ctx, msg.Digest())
 		require.True(t, found)
-		require.Equal(t, cctx.CctxStatus.Status, types.CctxStatus_OutboundMined)
+		require.Equal(t, types.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
 		require.Equal(t, cctx.InboundTxParams.TxFinalizationStatus, types.TxFinalizationStatus_Executed)
 	})
 
@@ -299,7 +308,8 @@ func TestStatus_ChangeStatus(t *testing.T) {
 
 func TestKeeper_SaveInbound(t *testing.T) {
 	t.Run("should save the cctx", func(t *testing.T) {
-		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
+		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
+		zk.ObserverKeeper.SetTSS(ctx, sample.Tss())
 		receiver := sample.EthAddress()
 		amount := big.NewInt(42)
 		senderChain := getValidEthChain(t)
@@ -313,7 +323,7 @@ func TestKeeper_SaveInbound(t *testing.T) {
 	})
 
 	t.Run("should save the cctx and remove tracker", func(t *testing.T) {
-		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
+		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		receiver := sample.EthAddress()
 		amount := big.NewInt(42)
 		senderChain := getValidEthChain(t)
@@ -326,6 +336,7 @@ func TestKeeper_SaveInbound(t *testing.T) {
 			CoinType: 0,
 		})
 		eventIndex := sample.Uint64InRange(1, 100)
+		zk.ObserverKeeper.SetTSS(ctx, sample.Tss())
 
 		k.SaveInbound(ctx, cctx, eventIndex)
 		require.Equal(t, types.TxFinalizationStatus_Executed, cctx.InboundTxParams.TxFinalizationStatus)